Data.Tuple.TH.Curry

Documentation

curry16 :: ((a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p) -> t) -> a -> b -> c -> d -> e -> f -> g -> h -> i -> j -> k -> l -> m -> n -> o -> p -> t

curry15 :: ((a, b, c, d, e, f, g, h, i, j, k, l, m, n, o) -> t) -> a -> b -> c -> d -> e -> f -> g -> h -> i -> j -> k -> l -> m -> n -> o -> t

curry14 :: ((a, b, c, d, e, f, g, h, i, j, k, l, m, n) -> t) -> a -> b -> c -> d -> e -> f -> g -> h -> i -> j -> k -> l -> m -> n -> t

curry13 :: ((a, b, c, d, e, f, g, h, i, j, k, l, m) -> t) -> a -> b -> c -> d -> e -> f -> g -> h -> i -> j -> k -> l -> m -> t

curry12 :: ((a, b, c, d, e, f, g, h, i, j, k, l) -> t) -> a -> b -> c -> d -> e -> f -> g -> h -> i -> j -> k -> l -> t

curry11 :: ((a, b, c, d, e, f, g, h, i, j, k) -> t) -> a -> b -> c -> d -> e -> f -> g -> h -> i -> j -> k -> t

curry10 :: ((a, b, c, d, e, f, g, h, i, j) -> t) -> a -> b -> c -> d -> e -> f -> g -> h -> i -> j -> t

curry9 :: ((a, b, c, d, e, f, g, h, i) -> t) -> a -> b -> c -> d -> e -> f -> g -> h -> i -> t

curry8 :: ((a, b, c, d, e, f, g, h) -> t) -> a -> b -> c -> d -> e -> f -> g -> h -> t

curry7 :: ((a, b, c, d, e, f, g) -> t) -> a -> b -> c -> d -> e -> f -> g -> t

curry6 :: ((a, b, c, d, e, f) -> t) -> a -> b -> c -> d -> e -> f -> t

curry5 :: ((a, b, c, d, e) -> t) -> a -> b -> c -> d -> e -> t

curry4 :: ((a, b, c, d) -> t) -> a -> b -> c -> d -> t

curry3 :: ((a, b, c) -> t) -> a -> b -> c -> t

curry2 :: ((a, b) -> t) -> a -> b -> t

curry1 :: (t1 -> t2) -> t1 -> t2

uncurry16 :: (t1 -> t2 -> t3 -> t4 -> t5 -> t6 -> t7 -> t8 -> t9 -> t10 -> t11 -> t12 -> t13 -> t14 -> t15 -> t16 -> t17) -> (t1, t2, t3, t4, t5, t6, t7, t8, t9, t10, t11, t12, t13, t14, t15, t16) -> t17

uncurry15 :: (t1 -> t2 -> t3 -> t4 -> t5 -> t6 -> t7 -> t8 -> t9 -> t10 -> t11 -> t12 -> t13 -> t14 -> t15 -> t16) -> (t1, t2, t3, t4, t5, t6, t7, t8, t9, t10, t11, t12, t13, t14, t15) -> t16

uncurry14 :: (t1 -> t2 -> t3 -> t4 -> t5 -> t6 -> t7 -> t8 -> t9 -> t10 -> t11 -> t12 -> t13 -> t14 -> t15) -> (t1, t2, t3, t4, t5, t6, t7, t8, t9, t10, t11, t12, t13, t14) -> t15

uncurry13 :: (t1 -> t2 -> t3 -> t4 -> t5 -> t6 -> t7 -> t8 -> t9 -> t10 -> t11 -> t12 -> t13 -> t14) -> (t1, t2, t3, t4, t5, t6, t7, t8, t9, t10, t11, t12, t13) -> t14

uncurry12 :: (t1 -> t2 -> t3 -> t4 -> t5 -> t6 -> t7 -> t8 -> t9 -> t10 -> t11 -> t12 -> t13) -> (t1, t2, t3, t4, t5, t6, t7, t8, t9, t10, t11, t12) -> t13

uncurry11 :: (t1 -> t2 -> t3 -> t4 -> t5 -> t6 -> t7 -> t8 -> t9 -> t10 -> t11 -> t12) -> (t1, t2, t3, t4, t5, t6, t7, t8, t9, t10, t11) -> t12

uncurry10 :: (t1 -> t2 -> t3 -> t4 -> t5 -> t6 -> t7 -> t8 -> t9 -> t10 -> t11) -> (t1, t2, t3, t4, t5, t6, t7, t8, t9, t10) -> t11

uncurry9 :: (t1 -> t2 -> t3 -> t4 -> t5 -> t6 -> t7 -> t8 -> t9 -> t10) -> (t1, t2, t3, t4, t5, t6, t7, t8, t9) -> t10

uncurry8 :: (t1 -> t2 -> t3 -> t4 -> t5 -> t6 -> t7 -> t8 -> t9) -> (t1, t2, t3, t4, t5, t6, t7, t8) -> t9

uncurry7 :: (t1 -> t2 -> t3 -> t4 -> t5 -> t6 -> t7 -> t8) -> (t1, t2, t3, t4, t5, t6, t7) -> t8

uncurry6 :: (t1 -> t2 -> t3 -> t4 -> t5 -> t6 -> t7) -> (t1, t2, t3, t4, t5, t6) -> t7

uncurry5 :: (t1 -> t2 -> t3 -> t4 -> t5 -> t6) -> (t1, t2, t3, t4, t5) -> t6

uncurry4 :: (t1 -> t2 -> t3 -> t4 -> t5) -> (t1, t2, t3, t4) -> t5

uncurry3 :: (t1 -> t2 -> t3 -> t4) -> (t1, t2, t3) -> t4

uncurry2 :: (t1 -> t2 -> t3) -> (t1, t2) -> t3

uncurry1 :: (t1 -> t2) -> t1 -> t2