Test.Expr.Reflection

Documentation

module Language.Haskell.TH

Assertion types

type ASTCheck = [Dec] -> Q ()

Any function with type ASTCheck declared in the teacher's file will be executed if reflection is enabled

Declaration checks

isComprehentionD :: Dec -> Q ()

Fails if a declaration is not a list comprehention