Test.Expr.Types

Description

Functions for working with Template Haskell type representation.

(c) 2018–2021 Vladimír Štill

Documentation

arity :: Type -> Int

uncurryType :: Type -> ([Type], Type)

Get a list of types of all arguments and the return type. All context are stripped.

isFunctionType :: Type -> Bool

Is the top-level type constructor a fully applied (->)?

hasInstance :: Type -> Name -> Q Bool

Does given type have instance of given class?

normalizeContext :: Type -> Type

Merge type contexts which can be merged. Note: this functions works correctly only if none of the quantifiers shadows any type variable. This seems to be the case both for reified types (where multiple contexts occur for class methods) and for [t| … |] expressions.

getTVars :: Type -> Set Name

Extract all type variables from a type.

regenTVars :: String -> Type -> Q Type

rewriteAnnotations :: Type -> Type

stripAnnotations :: Type -> Type

unannotate :: Type -> (Type, Type)

unannotate t ≈ (stripAnnotations t, rewriteAnnotations t)

type TestAs α β = α

Intended to be used with the teacher's solution: foo :: a `TestAs` Positive a -> b `TestAs` Large b -> Bool

Type Substitution

type Substitution = [(Name, Type)]

Type substitution. Please note that substitution is ordered, i.e. s1 = [(a, VarT b), (b, ConT Int)] is different from s2 = [(b, ConT Int), (a, VarT b)] as for a -> b // s1 produces Int -> Int while a -> b // s2 produces b -> Int.

substitute :: Substitution -> Type -> Type

(//) :: Type -> Substitution -> Type infixl 9

type Renaming = [(Name, Name)]

Renaming of type variables. Similar to Substitution, Renaming is ordered.

rename :: Renaming -> Type -> Type

Unification

data UniTypeId

Constructors

LeftType

RightType

BothTypes

data TypeOrder

Constructors

TUnifiable

TLessGeneral

TMoreGeneral

TEqual

unify :: Type -> Type -> Q (Either (UniTypeId, String) (TypeOrder, Type, Substitution))

Unify two types

There are certain limitations, i.e. some unifyable types will not be unified: * higher rank types (nested foralls) are not supported * explicit parenthesis in types are not supported * explicit kind signatures will work only it they are exactly matching (i.e. type variable with kind signature cannot be unified with a constructor or a variable without kind signature) * quantified type variables have proper kind signatures only if they had them in both of the original types * type contexts cannot be simplified, so we can get contexts which constraints such as Functor [] and types misjudged as TUnifiable instead of the actual ordering

unifyingSubstitution :: Type -> Type -> Either (UniTypeId, String) Substitution

Outputs unification of two types, the types must not share any type variables.

Printing

ppty :: Type -> String

Print a simplified type, more similar to the type printed by GHCi. There are some serious limitations: * the namespaces of types are stripped, regardless of its correctness * explicit foralls are strippend, regardless of need for kind signatures in them