Test.QuickCheck.Convertible

Description

Conversions for tests, this adds possiblity to covert lists of Fun wrapped functions to lists of functions. Note that the Convertible typeclass is pruposedly not exposed to avoid unexpected results.

Documentation

convert :: Convertible a b => a -> b