Test.QuickCheck.Literal

Description

Defines newtypes for strings and integers that can be used to generate constants. The concrete value of the constant is encoded in the type, e.g., 'StringConst "hi!"' will always generate string "hi!" using arbitrary. It is expected the values will be coerced to the underlying data type before use using coerce or fromConst. Converting to IntConst or StrConst using Coercible is undefined behaviour unless the actual value matches the value in the type.

Documentation

newtype IntConst (n :: Nat) (i :: Type)

The data constructor should not be used, can be only created by arbitrary or def

Constructors

IntConst i

newtype StrConst (sym :: Symbol) (str :: Type)

The data constructor should not be used, can be only created by arbitrary or def

Constructors

StrConst str

type StringConst (sym :: Symbol) = StrConst sym String

class FromConst c where

Associated Types

type ConstRep c :: Type

Methods

fromConst :: c -> ConstRep c