Test.Testable.Data

Description

Support for testing data type declarations.

(c) 2014 Vladimír Štill

Documentation

showData :: DataDecl -> String

Get string representation of DataDecl.

reflectData :: forall a. Data a => a -> DataDecl

infer data declaration of given type it works well for monomorphic ATDs, for polymorphic types it will print data declaration as it would seem if all type variables were instantiated. For records it ommits record fields For non-ADT types results will be wrong most likely

conArgs :: Data a => a -> [TypeRep]

Get types of constructor arguments of given value.

conArgsNames :: Data a => a -> [String]

Get string types of constructor arguments.

constrArgs :: forall witness. Data witness => witness -> Constr -> [TypeRep]

As conArgs but value is build from given type witness and constuctor representation.

constrArgsNames :: forall witness. Data witness => witness -> Constr -> [String]

As conArgsNames but value is build from given type witness and constuctor representation.

data DataDecl

Representation of (monomorphic) data type.

Constructors

DataDecl

Fields

dataDeclName :: String

dataDeclCons :: [DataCon]

data DataCon

Representatio of data constructor.

Constructors

DataCon

Fields

dataConName :: String

dataConFields :: [String]

canonizeData :: DataDecl -> DataDecl

Get canonical representation of data type, that is such that data constructor order is sorted.

deepCanonizeData :: DataDecl -> DataDecl

Canonize also data field order, in addition to data constructor order (as done by canonizeData).