Test.Testable.IO.Base

IO replacements

data IO a

IO is opaque, as in Prelude, but it does not interract with outer world therefore it is safe and escapable by means of runIOLines

Internally, we use ST to accumulate user output and line generator to represent input.

getLine :: IO String

Read a line from the standard input device.

getChar :: IO Char

Read a character from the standard input device.

getContents :: IO String

The getContents operation returns all user input as a single string, which is read lazily as it is needed

readIO :: Read a => String -> IO a

The readIO function is similar to read except that it signals parse failure to the IO monad instead of terminating the program. NOTE: as our IO has no exception support, this is pureIO . read

readLn :: Read a => IO a

The readLn function combines getLine and readIO.

putChar :: Char -> IO ()

Write a character to the standard output device.

putStr :: String -> IO ()

Write a string to the standard output device.

putStrLn :: String -> IO ()

The same as putStr, but adds a newline character.

print :: Show a => a -> IO ()

The print function outputs a value of any printable type to the standard output device. Printable types are those that are instances of class Show; print converts values to strings for output using the show operation and adds a newline.

For example, a program to print the first 20 integers and their powers of 2 could be written as:

main = print ([(n, 2^n) | n <- [0..19]])

running

runIOLines :: (Int -> String) -> IO a -> (a, [String])

Same as runIOLines', but does not allow finite input.

runIOLines' :: (Int -> Maybe String) -> IO a -> (a, [String])

Run virtual IO

This function accepts line genrator and (virtual) IO action to run

generator: a function which for given line number returns line of input. Lines are numbered from 0, each line must not contain newline ('\n') character, not even at the end.

action: a virtual IO action

return value is tuple (output of IO action, list of lines produced by the IO action). If no lines are produced then [""] returned. Also, empty line at the end is to be expected if '\n' was last character on output (for example if putStrLn was last action).

>>> runIOLines show $ return 0
(0, [""])
>>> runIOLines show $ getLine >>= print . (+1) . read >> getLine >>= \x -> return (read x :: Int)
(1, ["1",""])

functor, applicative and monad implementations

fmapIO :: (a -> b) -> IO a -> IO b

pureIO :: a -> IO a

appIO :: IO (a -> b) -> IO a -> IO b

bindIO :: IO a -> (a -> IO b) -> IO b