Test.Expr

Description

Simple utility functions for testing.

(c) 2014-2018 Vladimír Štill

Test Entry

testMain :: TestConfig -> Q [Dec]

testType :: ExprName -> Q Exp

extractOptionDef :: String -> Exp -> Q Exp

extractOptionMaybe :: String -> Q Exp

extractOption :: String -> Q Exp

Test Expression Building Blocks

(<==>) :: (Eq a, Show a, NFData a) => a -> a -> Property infix 4

Exception aware comparison. If no exception is thrown when evaluating either of the values, compares them using (==), if exception is thrown in both, exception type is compared, otherwise if exception is thrown only in one, property fails. Mismatching values are returned using counterexample on failure.

testArgs :: Args

runProperty :: Testable prp => Args -> prp -> IO ()

runProperties :: Testable prp => Args -> [prp] -> IO ()

data Args

Constructors

Args

Fields

replay :: Maybe (QCGen, Int)

maxSuccess :: Int

maxDiscardRatio :: Int

maxSize :: Int

chatty :: Bool

maxShrinks :: Int

scheduleAlarm :: Int -> IO Int

scheduleAlarm i calls alarm to schedule a real time alarm at least i seconds in the future.