Test.QuickCheck.Convertible

Description

Conversions for tests, this adds possiblity to covert lists of Fun wrapped functions to lists of functions. Note that you should never need to use the Convertible typeclass in you code and you cannot write new instances for it (to prevent unexpected behaviour in tests).

Documentation

class Convertible a b

Minimal complete definition

convert'

convert :: Convertible a b => a -> b