Test.QuickCheck.Range

Description

Entenstion of QuickCheck's modifiers with Integral ranges with type defined bounds.

>>> foo :: Range Int 0 255 -> Bool
>>> foo (Range v) = /* ... */

(c) 2014 Vladimír Štill

Documentation

type Range (i :: *) (from :: Nat) (to :: Nat) = Ranges i '['(from, to)]

Modifier for generating integral values within given inclusive range.

Range Int 0 42 will have arbitrary values of type Int in range from 0 to 42 inclusive.

newtype Ranges :: * -> [(Nat, Nat)] -> * where

Modifier for choosing arbitrarily from multiple ranges (first range is chosen uniformly, than value from this range is chosen uniformly).

For example, Ranges Int [(0,0), (10, 19)] will with probability 1/2 generate 0, and with probability 1/20 one of (10, 19) inclusive.

Constructors

Range :: {..} -> Ranges i ranges

Fields

unRange :: i

newtype CharRanges ranges

Constructors

CharRange (Ranges Char ranges)

unCharRange :: CharRanges ranges -> Char

charRangeToRange :: CharRanges ranges -> Ranges Char ranges

type CharRange (from :: Nat) (to :: Nat) = CharRanges '['(from, to)]

type AsciiPrintableRange = CharRange 32 126

class CRange (a :: k) where

Convert compile time type ranges to runtime values.

Minimal complete definition

toRanges

Methods

toRanges :: Proxy a -> [(Integer, Integer)]

inRanges :: Integral i => i -> [(Integer, Integer)] -> Bool

is given value in range?

newtype BoundedList (a :: *) (from :: Nat) (to :: Nat)

Constructors

BoundedList [a]

unBoundedList :: BoundedList a from to -> [a]